restructure modules, flake8 runs again
