+++ news_signals/technical_indicators.py
@@ -1,0 +1,19 @@
+
+
+def RSI(prices, interval=14):
+    """Computes Relative Strength Index given a price series and lookback interval
+    Modified from https://stackoverflow.com/questions/20526414/relative-strength-index-in-python-pandas
+    See more here https://www.investopedia.com/terms/r/rsi.asp"""
+    delta = prices.diff()
+
+    dUp, dDown = delta.copy(), delta.copy()
+    dUp[dUp < 0] = 0
+    dDown[dDown > 0] = 0
+
+    RolUp = dUp.rolling(interval).mean()
+    RolDown = dDown.rolling(interval).mean().abs()
+
+    RS = RolUp / RolDown
+    RSI = 100.0 - (100.0 / (1.0 + RS))
+    return RSI
+
